更新删帖逻辑

+++ src/main/java/net/gupt/community/mapper/CommonMapper.java
@@ -28,7 +28,7 @@ public interface CommonMapper {
      * @param id          帖子Id
      * @return int
      */
-    int deleteArticleByIdAndType(Integer articleType, Integer id);
+    int deleteArticleByIdAndType(Byte articleType, Integer id);
 
 
     /**
